🚀 Refactor: Change name from "config" to "configenvironment" for better clarity.
