+++ src/config/passport.config.js
@@ -1,7 +1,7 @@
 import passport from 'passport';
 import local from 'passport-local';
 import GitHubStrategy from 'passport-github2';
-import configObject from './config.js';
+import configObject from './configenvironment.js';
 import UserModel from '../model/user.models.js';
 import encrypt from '../helpers/encrypt.js';
 import CartModel from '../model/carts.models.js';
